Export registration e2e test in its own file
